extract python client models package - rk
