+++ client-modules/python/src/taskslock/client/models/task_lock.py
@@ -1,0 +1,10 @@
+
+from typing import Callable
+
+
+class TaskLock:
+    def __init__(self, api_response: dict, release: Callable):
+        self.task_name = api_response['taskName']
+        self.context_id = api_response['contextId']
+        self.isLocked = api_response['isLockAcquired']
+        self.release = release
