+++ client-modules/python/src/taskslock/client/tasks_lock_service.py
@@ -1,15 +1,6 @@
 from utils import host_utils
 import requests
 import json
-from typing import Callable
-
-
-class TaskLock:
-    def __init__(self, api_response: dict, release: Callable):
-        self.task_name = api_response['taskName']
-        self.context_id = api_response['contextId']
-        self.isLocked = api_response['isLockAcquired']
-        self.release = release
 
 
 class TasksLockService:
